% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abec_function.R
\name{abec}
\alias{abec}
\title{Area between exchange curves.}
\usage{
abec(proteinName, state0Name, state1Name, observations, timeList, peptidesFile,
  state0DFracFile, state1DFracFile, state0SDFile, state1SDFile)
}
\arguments{
\item{proteinName}{A String.}

\item{state0Name}{A String.}

\item{state1Name}{A String.}

\item{observations}{A Number or a List of numbers.}

\item{timeList}{A List of numbers. Each is the number of seconds a batch has been exposed to deuterium.}

\item{peptidesFile}{A String.}

\item{state0DFracFile}{A String.}

\item{state1DFracFile}{A String.}

\item{state0SDFile}{A String.}

\item{state1SDFile}{A String.}
}
\value{
myOutput         meanX, meanY, DifMeans, Sdm, pvaluedm, Areabc, Sbc, pvaluebc
         meanX           mean remaining hydrogen fractions of state 0.
         meanY           mean remaining hydrogen fractions of state 1.
         DifMeans        Difference of the mean hydrogen remaining between the two states.
         Sdm             The standard deviation of the difference of means.
         pvaluedm        The p-value of the difference of means.
         Areabc          Area between the exchange curves.
         Sbc             Standard deviation of the curve area.
         pvaluebc        The p-value of the area between the curves.
}
\description{
Area between exchange curves.
}
\examples{
abec("PPARg", "apo", "Rosi", 4, c(1, 30, 60, 900, 3600), "Peptides.txt",
 "data/apo.txt", "data/Rosi.txt", "data/apoSD.txt", "data/RosiSD.txt")
}

